trying to link all the entities and their methods

+++ src/main/java/es/swapsounds/controller/SoundController.java
@@ -132,7 +132,7 @@ public class SoundController {
         String imagePath = storage.saveFile(username, imageFile, "images");
 
         // Creating and stroring the sound
-        Sound sound = new Sound(0, title, description, audioPath, imagePath, userId, category, duration);
+        Sound sound = new Sound((long)0, title, description, audioPath, imagePath, userId, category, duration);
         storage.addSound(sound);
 
         model.addAttribute("success", "Sound uploaded successfully!");
